Feature: Pass widget draw_slider instead of values
Previously, I was passing the values contained in widgets to the slider
instead of just passing the object itself. As I need to have the sliders
have access to more and more variables, this becomes less and less
useful. I also will want the sliders to be able to report their size
back to the widget struct, so I need the reference.