Feature: Determine widget ID using bounding boxes
If the mouse has been clicked, we want to know which widget was clicked
so that we can only apply further operations to that widget. This is
important if the user clicks on widget 0, but drags their mouse into
widget 1. We still want to be operating on widget 0 in that case.