Feature: Propagate render windows through code
I've created the concept of "contexts". Contexts are essentially
rectangles that will allow us to have some primitive camera controls.

In this case, the parent and child contexts are equivalent, but in
principle you could have something like this:

 .----------------------------------------.
 |                                        |
 |  .----------------.  Parent context    |
 |  |                |                    |
 |  | Child context  |                    |
 |  |                |                    |
 |  |                |                    |
 |  '----------------'                    |
 |                                        |
 |                                        |
 |                                        |
 '----------------------------------------'

The rendering code can then take everything in the child context and
apply a transformation to it into the parent context, essentially
"zooming in".

This concept will become more clear in future commits.